Change test-project naming
