+++ source/Databricks/source/SqlStatementExecution.UnitTests/Assets/CalculationResult.json
