+++ source/Databricks/source/SqlStatementExecution.UnitTests/Assets/CalculationResultChunk.json
