+++ source/Databricks/source/SqlStatementExecution.UnitTests/Assets/CalculationResultWithExternalLinks.json
