+++ source/Databricks/source/SqlStatementExecution.UnitTests/EmbeddedResources.cs
@@ -13,9 +13,9 @@
 // limitations under the License.
 
 using System.Reflection;
-using Energinet.DataHub.Core.Databricks.SqlStatementExecution.Tests.Assets;
+using Energinet.DataHub.Core.Databricks.SqlStatementExecution.UnitTests.Assets;
 
-namespace Energinet.DataHub.Core.Databricks.SqlStatementExecution.Tests;
+namespace Energinet.DataHub.Core.Databricks.SqlStatementExecution.UnitTests;
 
 // TODO BJM: Copied from other project. Can we share instead? Possibly in Test.Core?
 public static class EmbeddedResources
