+++ source/Databricks/source/SqlStatementExecution.UnitTests/Internal/DatabricksSqlStatementClientTests.cs
@@ -16,12 +16,12 @@ using System.Net;
 using Energinet.DataHub.Core.Databricks.SqlStatementExecution.Abstractions;
 using Energinet.DataHub.Core.Databricks.SqlStatementExecution.Internal;
 using Energinet.DataHub.Core.Databricks.SqlStatementExecution.Models;
-using Energinet.DataHub.Core.Databricks.SqlStatementExecution.Tests.Builders;
+using Energinet.DataHub.Core.Databricks.SqlStatementExecution.UnitTests.Builders;
 using Energinet.DataHub.Core.TestCommon.AutoFixture.Attributes;
 using Microsoft.Extensions.Logging;
 using Moq;
 
-namespace Energinet.DataHub.Core.Databricks.SqlStatementExecution.Tests.Internal;
+namespace Energinet.DataHub.Core.Databricks.SqlStatementExecution.UnitTests.Internal;
 
 public class DatabricksSqlStatementClientTests
 {
