+++ source/Databricks/source/SqlStatementExecution.UnitTests/SqlResponseStatusHelper.cs
@@ -12,8 +12,19 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace Energinet.DataHub.Core.Databricks.SqlStatementExecution.Tests.Assets;
+using Newtonsoft.Json;
 
-internal class Root
+namespace Energinet.DataHub.Core.Databricks.SqlStatementExecution.UnitTests;
+
+public static class SqlResponseStatusHelper
 {
+    public static string CreateStatusResponse(string state)
+    {
+        var statement = new
+        {
+            statement_id = "01edef23-0d2c-10dd-879b-26b5e97b3796",
+            status = new { state, },
+        };
+        return JsonConvert.SerializeObject(statement, Formatting.Indented);
+    }
 }
